python/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-framework"
version = "0.1.0"
description = "Enterprise agentic framework - Python SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
redis = ["redis>=5.0"]
openai = ["openai>=1.0"]
observability = ["opentelemetry-api>=1.20"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
